Moved files into roslaunch and placed __init__.py for the new modules
