+++ setup.cfg
@@ -60,10 +60,12 @@ per-file-ignores =
   src/roswire/proxy/roslaunch/config/node.py:N817
   src/roswire/proxy/roslaunch/config/launch.py:N817
   src/roswire/proxy/roslaunch/config/parameter.py:N817
-  src/roswire/proxy/roslaunch/reader.py:F811,E704,E501,N817
+  src/roswire/ros1/roslaunch/reader.py:F811,E704,E501,N817
   src/roswire/proxy/roslaunch/roslaunch.py:N817
   src/roswire/ros1/__init__.py:F401
   src/roswire/ros2/__init__.py:F401
+  src/roswire/ros1/roslaunch/__init__.py:F401
+  src/roswire/ros2/roslaunch/__init__.py:F401
   src/roswire/service/__init__.py:F401
 
 [tox]
