+++ src/roswire/ros1/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 from .manager import ROS1NodeManager
 from .node import ROS1Node
-from .roslaunch import ROS1LaunchManager
+from .roslaunch.roslaunch import ROS1LaunchManager
