+++ src/roswire/ros1/roslaunch/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .roslaunch import ROS1LaunchManager
