+++ src/roswire/ros2/roslaunch/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .launch import ROS2LaunchManager
