+++ src/roswire/ros2/roslaunch/reader.py
@@ -9,11 +9,11 @@ import attr
 import pkg_resources
 from loguru import logger
 
-from ..proxy.roslaunch.config import ExecutableType, LaunchConfig, NodeConfig
-from ..proxy.roslaunch.reader import LaunchFileReader
+from ...proxy.roslaunch.config import ExecutableType, LaunchConfig, NodeConfig
+from ...proxy.roslaunch.reader import LaunchFileReader
 
 if typing.TYPE_CHECKING:
-    from .. import AppInstance
+    from roswire import AppInstance
 
 
 @attr.s(auto_attribs=True)
